cmd: refine config and disable wip features
